requirements.txt
numpy
fastapi
uvicorn
pydantic
pydantic-settings
prometheus_client
loguru
python-multipart
requests
google-generativeai
pytest
httpx
python-dotenv
psutil
redis
slowapi
scipy
pandas
tenacity
